drop udisks2 from server profile
As far as I can see only desktop environments are enabling this option.
This is a rather low-level service, which seems unlikely to get enabled
by accident.
However setting this option to false will break `xrdp` in combination
with a desktop environment usage which is something you may need every
once in a while even in a server context.