add _file self-references to imported modules
I wish there was a better option.

The _file reference is used by the module evaluations to give better
context when bad options are being set. Normally, this reference is
added recursively during the module import.

Because `nix flake check` throws an error if modules are not a function
or a set, we are forced to use the `import` statement and lose that
context. So we annotate the files directly instead.