flake: remove flake-parts dependency
One drawback of flake-parts is that it makes the wiring of the flake
outputs less easy to follow. With the new version the user can open the
flake.nix and pretty easily follow how things are put together.

Another drawback is that it makes the flake dependency tree larger. We
want to limit how much baggage srvos is adding when added as an input.