hetzner-online: use predictable interface names
...and match all ethernet devices according to https://www.freedesktop.org/wiki/Software/systemd/PredictableNetworkInterfaceNames/

We still assume that the machine either has just one interface
or that it's at least harmless to send dhcp requests on all of them

matchConfig.Name is a whitespace-separated list of shell-style globs according
to man 5 systemd.network.

old-style eth0 is kept to ease switching between interface naming
schemes and because it seems harmless to do so.