mdns: enable for all useDHCP interfaces
`nixos/modules/tasks/network-interfaces-systemd.nix` only creates
networkd networks if `useDHCP` is enabled. This is the case for the
global setting, as well as the per-interface setting.

This commit only enables the multicast option globally if `useDHCP` is
enabled globally. Additionally it enables MulticastDNS for each
interface that has `useDHCP` enabled.

Signed-off-by: Sefa Eyeoglu <contact@scrumplex.net>