feat: add hostname change detection
NixOS is just too easy to deploy to the wrong host. Eg:

    nixos-rebuild --flake .#wronghostname --target-host root@otherhost switch

It happened to me a few times, and when I started talking about it I
got the same feedback from other people.

After this change, the user now has to confirm with YES before the
switch can happen.

To disable, set

    srvos.detect-hostname-change.enable = false;